+++ src/app/locales/en/utils/file-info.js
@@ -1,3 +1,4 @@
+import sizeTabModel from './file-info/size-tab-model';
 import metadataTabModel from './file-info/metadata-tab-model';
 import permissionsTabModel from './file-info/permissions-tab-model';
 import sharesTabModel from './file-info/shares-tab-model';
@@ -5,6 +6,7 @@ import qosTabModel from './file-info/qos-tab-model';
 import distributionTabModel from './file-info/distribution-tab-model';
 
 export default {
+  sizeTabModel,
   metadataTabModel,
   permissionsTabModel,
   sharesTabModel,